refactor(control-plane): extract prompt enqueue contract
