+++ packages/control-plane/src/session/services/message.service.ts
@@ -1,36 +1,11 @@
 import type { ArtifactRow } from "../types";
-import { sessionAttachmentReferencesSchema, type SessionMessage } from "@open-inspect/shared";
+import type { SessionMessage } from "@open-inspect/shared";
 import type { ArtifactResponse, ListEventsResponse } from "../../types";
 import type { SessionRepository } from "../repository";
 import type { SessionMessageQueue } from "../message-queue";
+import type { EnqueuePromptRequest } from "../enqueue-prompt-contract";
 import { SessionEventStream, type SessionEventListRequest } from "../event-stream";
 import { parseStoredSessionAttachments } from "../session-attachment-resolver";
-import { z } from "zod";
-
-export const enqueuePromptRequestSchema = z.object({
-  content: z.string(),
-  authorId: z.string(),
-  canonicalUserId: z.string().nullable().optional(),
-  source: z.string(),
-  model: z.string().optional(),
-  reasoningEffort: z.string().optional(),
-  attachments: sessionAttachmentReferencesSchema.optional(),
-  callbackContext: z.record(z.string(), z.unknown()).optional(),
-  // Trusted SCM enrichment resolved by the router at prompt time.
-  scmEnrichment: z
-    .object({
-      userId: z.string().nullable(),
-      login: z.string().nullable(),
-      name: z.string().nullable(),
-      email: z.string().nullable(),
-      accessTokenEncrypted: z.string().nullable(),
-      refreshTokenEncrypted: z.string().nullable(),
-      tokenExpiresAt: z.number().nullable(),
-    })
-    .optional(),
-});
-
-export type EnqueuePromptRequest = z.infer<typeof enqueuePromptRequestSchema>;
 
 export type ListEventsRequest = SessionEventListRequest;
